Refactored the ued script into a class called MicroED
