+++ pyTEM/lib/micro_ed/ico/BASF.ico
